psl: Ensure `set_config_dir` can be eliminated if not used
Due to dynamic dispatch, optimizer does not remove it from WASM bundle,
even though it is not used. It brings in heavy URL parser that adds
around 270K uncompressed/130K gzip. This PR rewrites it to be a free
function outside of `Connector` trait that does not have this problem.

Fix prisma/team-orm#772
