+++ psl/psl-core/Cargo.toml
@@ -11,6 +11,7 @@ schema-ast = { path = "../schema-ast" }
 
 bigdecimal = "0.3"
 chrono = { version = "0.4.6", default_features = false }
+connection-string.workspace = true
 itertools.workspace = true
 once_cell = "1.3.1"
 regex = "1.3.7"
