+++ psl/psl-core/src/datamodel_connector.rs
@@ -226,49 +226,6 @@ pub trait Connector: Send + Sync {
         diagnostics: &mut Diagnostics,
     ) -> Option<NativeTypeInstance>;
 
-    fn set_config_dir<'a>(&self, config_dir: &std::path::Path, url: &'a str) -> Cow<'a, str> {
-        let set_root = |path: &str| {
-            let path = std::path::Path::new(path);
-
-            if path.is_relative() {
-                Some(config_dir.join(path).to_str().map(ToString::to_string).unwrap())
-            } else {
-                None
-            }
-        };
-
-        let mut url = match url::Url::parse(url) {
-            Ok(url) => url,
-            Err(_) => return Cow::from(url), // bail
-        };
-
-        let mut params: BTreeMap<String, String> =
-            url.query_pairs().map(|(k, v)| (k.to_string(), v.to_string())).collect();
-
-        url.query_pairs_mut().clear();
-
-        // Only for PostgreSQL + MySQL
-        if let Some(path) = params.get("sslcert").map(|s| s.as_str()).and_then(set_root) {
-            params.insert("sslcert".into(), path);
-        }
-
-        // Only for PostgreSQL + MySQL
-        if let Some(path) = params.get("sslidentity").map(|s| s.as_str()).and_then(set_root) {
-            params.insert("sslidentity".into(), path);
-        }
-
-        // Only for MongoDB
-        if let Some(path) = params.get("tlsCAFile").map(|s| s.as_str()).and_then(set_root) {
-            params.insert("tlsCAFile".into(), path);
-        }
-
-        for (k, v) in params.into_iter() {
-            url.query_pairs_mut().append_pair(&k, &v);
-        }
-
-        url.to_string().into()
-    }
-
     fn supports_scalar_lists(&self) -> bool {
         self.has_capability(ConnectorCapability::ScalarLists)
     }
